setup(ProductSlotsReelsContext): Dedicated types folder/file for context
