+++ context/ProductSlotsReelsContext/ProductSlotsReelsContext.tsx
@@ -15,26 +15,10 @@ import {
   CELL_HEIGHT,
   VISIBLE,
 } from "@/components/Slots/SlotsGame/components/ProductSlotsReels/enum";
-
-export type ProductSlotsReelsContextType = {
-  reels: Product[][];
-  pos: number[];
-  transitionDurations: number[];
-  animating: boolean[];
-  activePlaceholder: boolean;
-  spinning: boolean;
-  CELL_HEIGHT: number;
-  VISIBLE: number;
-};
+import { ProductSlotsReelsContextType, ProviderProps } from "./types";
 
 const ProductSlotsReelsContext =
   createContext<ProductSlotsReelsContextType | null>(null);
-
-type ProviderProps = {
-  selectedProducts: Product[];
-  spinning: boolean;
-  children: ReactNode;
-};
 
 export const ProductSlotsReelsProvider: React.FC<ProviderProps> = ({
   selectedProducts,
